requirements.txt
pypdf>=3.0
pdfminer.six>=20201018
tqdm
